Improve flags management (issue #102)
Context:
When adding flags to dice-roller inside a Table, the pipe
needs to be escaped not to be seen as a column separator by
the table. So flags can be found either as pipe or as escaped pipe
based on where the dice-roller is defined.

Changes:

TableRoller splits rows using pipe as the markdown column separator.
But pipe is also used to define flags for dice-rolls but, in this
case, they are escaped.
As negative lookbehind regex is noty supported by WebKit, split is
done in three steps:
    - replace escaped pipes by a temporary fixed text (without any pipe)
    - split using pipe
    - replace the fixed text by a pipe (no need to escape it anymore)

This now supports multiple flags defined for a given dice-roller.

Improve getRoller to manage multiple flags too (i.e. replacing not
only one escaped pipe by a pipe but all of them).